requirements.txt
flake8==3.5.0
gitdb2==2.0.4
GitPython==2.1.11
mccabe==0.6.1
mypy==0.630
mypy-extensions==0.4.1
pycodestyle==2.3.1
pyflakes==1.6.0
smmap2==2.0.4
SQLAlchemy==1.2.12
typed-ast==1.1.0
